requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-ipware==3.0.2
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pipenv==2020.11.15
psycopg2==2.8.6
PyJWT==2.0.1
pylint==2.7.3
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
